Add files via upload
# Land-Record-Management-Using-Hyperledger-Fabirc
In a world that is continuously innovating and evolving, Blockchain technology has immensely contributed to the transaction and storage methods which not only eliminates the manipulation of a trusted third party but also provides a distributed shared ledger. 
Buying and selling of a land (Land Registration) consists or number of steps which are complicated. Registration, the transfer of property ownership, and the declaration certificates are all examples of LRMS processes.
information is stored and distributed across various government organizations that are not well-synchronized in existing systems, which means a third party can easily interrupt it.
By using decentralized system, it will be easy to provide transparency, and no one has unfair power.
It will be a system where seller can upload its property to sell through registrar who will upload the property, relevant information about the property, and buyer can send the buy request if he is willing to buy.
Before a ownership is transferred, the buyer will have to deposit the stamp duties and clear all the dues and fees which are necessary in a smart contract. After the mutation certificate is transferred, the property will be transferred to the new owner whose information will be updated by registrar and uploaded on ledger.
# Installation
```bash
git clone https://github.com/Ammarkhalid047/Land-Record-Management-Using-Hyperledger-Fabirc.git
```
```bash
cd Land-Record-Management-Using-Hyperledger-Fabirc
```
```bash
cd landregistry
```
```bash
./startfabric.sh
```
```bash
cd javascript
```
```bash
npm install
```
```bash
npm start
```
## Contributing
Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.

Please make sure to update tests as appropriate.